Try some package refactoring to make capacity related jmx apis available in capacity-ingress deployment
