+++ src/main/java/org/candlepin/subscriptions/jmx/capacity/SubscriptionJmxBean.java
@@ -18,9 +18,8 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package org.candlepin.subscriptions.jmx;
+package org.candlepin.subscriptions.jmx.capacity;
 
-import javax.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
 import org.candlepin.subscriptions.db.model.OrgConfigRepository;
 import org.candlepin.subscriptions.resource.ResourceUtils;
@@ -29,6 +28,8 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.jmx.export.annotation.ManagedOperation;
 import org.springframework.jmx.export.annotation.ManagedResource;
 import org.springframework.stereotype.Component;
+
+import javax.transaction.Transactional;
 
 @Profile("capacity-ingress")
 @Component
